requirements.txt
requests
pandas
pyyaml
pycryptodome
websockets
fastapi
uvicorn
jinja2
pytest
itsdangerous
